fix: bug of ingesting table when there are multiple layers under a raw dataset
