+++ sites/geohub/src/routes/(app)/data/[id]/+page.svelte
@@ -33,19 +33,16 @@
 	<title>{title}</title>
 	<meta property="og:site_name" content={SiteInfo.site_name} />
 	<meta property="og:type" content="article" />
-	<meta name="description" content={SiteInfo.site_description} />
-	<meta property="og:description" content={SiteInfo.site_description} />
-	<meta name="twitter:description" content={SiteInfo.site_description} />
+	<meta name="description" {content} />
+	<meta property="og:description" {content} />
+	<meta name="twitter:description" {content} />
 	<meta property="og:title" content={title} />
-	<meta property="og:image" content="{$page.url.origin}/api/og?content={content}" />
+	<meta property="og:image" content="{$page.url.origin}/api/og?content={title}" />
 	<meta property="og:image:width" content="1200" />
 	<meta property="og:image:height" content="630" />
 	<meta name="twitter:card" content="summary_large_image" />
 	<meta name="twitter:title" content={title} />
-	<meta
-		name="twitter:image"
-		content="{$page.url.origin}/api/og?content={encodeURIComponent(content)}"
-	/>
+	<meta name="twitter:image" content="{$page.url.origin}/api/og?content={title}" />
 	<meta property="og:url" content="{$page.url.origin}{$page.url.pathname}" />
 </svelte:head>
